+++ packages/plait/src/transforms/selection.ts
@@ -1,6 +1,8 @@
 import { SetSelectionOperation } from '../interfaces/operation';
 import { PlaitBoard } from '../interfaces/board';
 import { Selection } from '../interfaces/selection';
+import { BOARD_TO_TEMPORARY_ELEMENTS } from '../utils/weak-maps';
+import { PlaitElement } from '../interfaces/element';
 
 export function setSelection(board: PlaitBoard, selection: Selection | null) {
     const operation: SetSelectionOperation = { type: 'set_selection', properties: board.selection, newProperties: selection };
@@ -9,8 +11,18 @@ export function setSelection(board: PlaitBoard, selection: Selection | null) {
 
 export interface SelectionTransforms {
     setSelection: (board: PlaitBoard, selection: Selection | null) => void;
+    setSelectionWithTemporaryElements: (board: PlaitBoard, elements: PlaitElement[]) => void;
 }
 
 export const SelectionTransforms: SelectionTransforms = {
-    setSelection
+    setSelection,
+    setSelectionWithTemporaryElements
 };
+
+
+export function setSelectionWithTemporaryElements(board: PlaitBoard, elements: PlaitElement[]) {
+    setTimeout(() => {
+        BOARD_TO_TEMPORARY_ELEMENTS.set(board, elements);
+        setSelection(board, { ranges: [] });
+    });
+}